requirements.txt
PyQt5
numpy
Pillow
PyOpenGL
pygame
PyGLM